coding panel will now instantiate slot based on dragged item size
