+++ client/CS130_Team13/Assets/Scenes/JohnnyWorkspace.unity
@@ -156,8 +156,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 130.51, y: -129}
-  m_SizeDelta: {x: 148.48824, y: 20}
+  m_AnchoredPosition: {x: 123, y: -196.2}
+  m_SizeDelta: {x: 148.48824, y: 34.401543}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &226374421
 MonoBehaviour:
@@ -212,6 +212,18 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 226374419}
   m_CullTransparentMesh: 0
+--- !u!114 &342579494 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3602615441031542704, guid: 773fd189e5c1c50489eed1f93821a1a9,
+    type: 3}
+  m_PrefabInstance: {fileID: 3602615439956202193}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0b8e49ee328a5bf418b6df60a61b3597, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &588423618
 GameObject:
   m_ObjectHideFlags: 0
@@ -247,8 +259,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 130.51, y: -129}
-  m_SizeDelta: {x: 148.48824, y: 20}
+  m_AnchoredPosition: {x: 130.51, y: -59.9}
+  m_SizeDelta: {x: 148.48824, y: 26.172104}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &588423620
 MonoBehaviour:
@@ -312,9 +324,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 997477689}
-  - component: {fileID: 997477690}
   m_Layer: 5
-  m_Name: SomeContainerThatMakesTheSceneComplex
+  m_Name: rThatMakesTheSceneComplexSceneComplex
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -340,22 +351,9 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0.000015795, y: 0.000000089407}
+  m_AnchoredPosition: {x: 0.000015795, y: 0}
   m_SizeDelta: {x: 324.03638, y: 134.75531}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &997477690
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 997477688}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 30ede02b0dd1b5f4fb961c0f006bbb5f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  myParentTransform: {fileID: 1990610671}
 --- !u!1 &1087432412
 GameObject:
   m_ObjectHideFlags: 0
@@ -391,8 +389,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 130.51, y: -129}
-  m_SizeDelta: {x: 148.48824, y: 20}
+  m_AnchoredPosition: {x: 68, y: -161.7}
+  m_SizeDelta: {x: 148.48824, y: 8.684509}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1087432414
 MonoBehaviour:
@@ -447,6 +445,28 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1087432412}
   m_CullTransparentMesh: 0
+--- !u!1 &1126826023 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 3602615440832939433, guid: 773fd189e5c1c50489eed1f93821a1a9,
+    type: 3}
+  m_PrefabInstance: {fileID: 3602615439956202193}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1126826025
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1126826023}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8105c104998dd1347ad1f42427b444f6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  myGuard: {fileID: 342579494}
+  mySlotInstance: {fileID: 174922305299236426, guid: 355244f3b8a42e74fbc3a6c4fd3b69fd,
+    type: 3}
+  mySlots: []
 --- !u!1 &1153391397
 GameObject:
   m_ObjectHideFlags: 0
@@ -548,50 +568,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!1 &1454374117
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1454374119}
-  - component: {fileID: 1454374118}
-  m_Layer: 0
-  m_Name: Drag Drop Manager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &1454374118
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1454374117}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ea8f19936a8b5d24592e47086a683557, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  containerDuringDragging: {fileID: 1370847711}
---- !u!4 &1454374119
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1454374117}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 348.59805, y: 209.36682, z: -111.749985}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 3
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1762919573
 GameObject:
   m_ObjectHideFlags: 0
@@ -627,8 +603,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 130.51, y: -129}
-  m_SizeDelta: {x: 148.48824, y: 20}
+  m_AnchoredPosition: {x: 117, y: -116.8}
+  m_SizeDelta: {x: 148.48824, y: 41.602303}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1762919575
 MonoBehaviour:
@@ -990,6 +966,36 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1990610675}
     m_Modifications:
+    - target: {fileID: 3602615440832939432, guid: 773fd189e5c1c50489eed1f93821a1a9,
+        type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: -200
+      objectReference: {fileID: 0}
+    - target: {fileID: 3602615440832939432, guid: 773fd189e5c1c50489eed1f93821a1a9,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0.024383
+      objectReference: {fileID: 0}
+    - target: {fileID: 3602615440832939436, guid: 773fd189e5c1c50489eed1f93821a1a9,
+        type: 3}
+      propertyPath: m_ChildAlignment
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3602615441031542704, guid: 773fd189e5c1c50489eed1f93821a1a9,
+        type: 3}
+      propertyPath: myPanel
+      value: 
+      objectReference: {fileID: 1126826025}
+    - target: {fileID: 3602615441031542705, guid: 773fd189e5c1c50489eed1f93821a1a9,
+        type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: -200
+      objectReference: {fileID: 0}
+    - target: {fileID: 3602615441031542705, guid: 773fd189e5c1c50489eed1f93821a1a9,
+        type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: -0.00020599
+      objectReference: {fileID: 0}
     - target: {fileID: 5434661751078821699, guid: 773fd189e5c1c50489eed1f93821a1a9,
         type: 3}
       propertyPath: m_Pivot.x
@@ -1100,7 +1106,13 @@ PrefabInstance:
       propertyPath: m_Name
       value: Coding Panel
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    - target: {fileID: 7969446061602245806, guid: 773fd189e5c1c50489eed1f93821a1a9,
+        type: 3}
+      propertyPath: containerDuringDragging
+      value: 
+      objectReference: {fileID: 1370847711}
+    m_RemovedComponents:
+    - {fileID: 3602615440832939435, guid: 773fd189e5c1c50489eed1f93821a1a9, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 773fd189e5c1c50489eed1f93821a1a9, type: 3}
 --- !u!224 &3602615439956202194 stripped
 RectTransform:
