+++ client/CS130_Team13/Assets/Scripts/Drag n Drop/DragDropManager.cs
@@ -6,7 +6,15 @@ public class DragDropManager : MonoBehaviour
 {
     public static DragDropManager instance = null;
 
+    /// <summary>
+    /// aux object that forces the dragged item to be displayed on top
+    /// </summary>
     [SerializeField] private GameObject containerDuringDragging = null;
+
+    /// <summary>
+    /// allows other objects to have access to the dragged item
+    /// </summary>
+    public GameObject currentlyDraggedItem = null;
 
     private void Awake() {
         // singleton
