+++ client/CS130_Team13/Assets/Scripts/Drag n Drop/Prototypes/Prototype Coding Panel/PanelItem.cs
@@ -1,0 +1,27 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.EventSystems;
+
+public class PanelItem : ProtoDraggable, ICodeInfo
+{
+    [SerializeField]
+    private string myInformation = "";
+    
+    public string GetInformation() {
+        return myInformation;
+    }
+
+    public override void OnPointerDown(PointerEventData eventData) {
+        base.OnPointerDown(eventData);
+
+        DragDropManager.instance.currentlyDraggedItem = gameObject;
+    }
+
+    public override void OnPointerUp(PointerEventData eventData) {
+        base.OnPointerUp(eventData);
+
+        if (DragDropManager.instance.currentlyDraggedItem == gameObject)
+            DragDropManager.instance.currentlyDraggedItem = null;
+    }
+}
