--- client/CS130_Team13/Assets/Scripts/Drag n Drop/Prototypes/Prototype Coding Panel/ProtoPanelPlug.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ProtoPanelPlug : ProtoDraggable, ICodeInfo
-{
-    [SerializeField]
-    private string myInformation = "";
-    
-    public string GetInformation() {
-        return myInformation;
-    }
-}
