+++ client/CS130_Team13/Assets/Scripts/Drag n Drop/Prototypes/Prototype Coding Panel/ProtoPanelSlot.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class ProtoPanelSlot : ProtoDroppable, ICodeInfo
 {
-    public ProtoCodingPanel myPanel;
+    public CodingPanel myPanel;
 
     private string myInfo = "";
